+++ plugin_scripts/__init__.py
@@ -1,5 +1,4 @@
 from .deploy import main
 
-
 if __name__ == "__main__":
     main()
