+++ plugin_scripts/deploy.py
@@ -4,7 +4,7 @@ import sys
 
 from gbq import BigQuery
 
-from .pipeline_exceptions import DeployFailed, DatasetSchemaDirectoryNonExistent
+from .pipeline_exceptions import DatasetSchemaDirectoryNonExistent, DeployFailed
 
 sys.tracebacklimit = 0
