+++ tests/test_deploy.py
@@ -87,28 +87,3 @@ def test_main_true(mocker, gcp_project, dataset_schema_directory, credentials):
     mocker.patch("plugin_scripts.deploy.BigQuery")
     deploy.main()
     assert True
-
-
-@pytest.mark.parametrize(
-    "exception_object",
-    [
-        DatasetSchemaDirectoryNonExistent({}),
-        DeployFailed({}),
-    ],
-)
-def test_exception_init(exception_object):
-    e = exception_object
-    assert isinstance(e, Exception)
-
-
-@pytest.mark.parametrize(
-    "exception_class",
-    [
-        DatasetSchemaDirectoryNonExistent,
-        DeployFailed,
-    ],
-)
-def test_exception_throws(exception_class):
-    e = exception_class({})
-    with pytest.raises(exception_class):
-        raise e
